refactor: split Map component into Map and GameMap
Extract the basic Map part of the Map component, and then have the new
GameMap component be for the actual game. This is so that a map can be
displayed on other parts of the website, without the game functionality.
