+++ src/app/types/BaseMapProps.ts
@@ -1,0 +1,10 @@
+import MapType from '@typings/MapType';
+
+interface BaseMapProps {
+    mapType: MapType;
+    className?: string;
+    minimized: boolean;
+    setMinimized: (value: boolean) => void;
+}
+
+export default BaseMapProps;
