+++ src/app/types/GameMapProps.ts
@@ -1,0 +1,6 @@
+import BaseMapProps from '@typings/BaseMapProps';
+
+interface GameMapProps extends BaseMapProps {
+}
+
+export default GameMapProps;
